fix(web/shipping): wrap the failed-shipment reason on mobile cards + lock the parity (#1800)
The #1800 status cell renders through a shared `ShipmentStatusCell` used by
BOTH the desktop table cell and the DataTable mobile card `meta` slot.

Production fix (UX): the persisted rejection reason was `white-space: nowrap`
+ ellipsis + `max-width: 32ch` with the full text only on native `title`
hover. That undercuts triage-from-phone (frontend-ui-style-guide § Responsive):
touch has no hover, so on the mobile card the operator saw only a clamped
fragment and had to open the order to read why. A card has vertical room, so
its reason now wraps to full text (`overflow-wrap: anywhere` keeps a long
unbroken token from overflowing). The dense desktop table row keeps the
one-line truncation + `title` (density budget; consistent with the codebase's
existing truncated-cell convention — no bespoke Tooltip primitive for a
table-cell string).

Test (regression): a mobile-viewport test (forces `useMediaQuery` via a
matchMedia mock) asserts the reason renders inside `.data-table__card`, so the
card path can't silently regress. jsdom can't assert CSS wrapping, so the test
locks the reason's *presence* on the card; the CSS above makes it *readable*.

Shared test helper: extract the matchMedia mobile-viewport mock into a
provider-free `test/viewport.ts#mockMobileViewport` (kept out of the heavier
`test-utils.tsx` so the `shared/ui` DataTable spec doesn't pull the app
provider/plugin harness into its graph); migrate `data-table.test.tsx`'s local
copy to it.

Follow-ups (list-triage recovery action + carrier-message role visibility)
tracked in #1826.

Docs: None — a responsive CSS fix to an existing surface + test hardening; no
new port, capability, contract, token, or UI pattern.
