+++ apps/web/src/index.css
@@ -11086,6 +11086,23 @@ textarea.bulk-editor__input { resize: vertical; min-height: 66px; font-family: i
   color: var(--text-muted);
 }
 
+/* #1800 — mobile card view. The dense desktop table row keeps the reason on
+   one truncated line (density budget) with the full text on native `title`
+   hover; a card has vertical room and no hover on touch, so let the reason
+   wrap to full text there — otherwise triage-from-phone can only read a
+   clamped fragment. `overflow-wrap: anywhere` keeps a long unbroken token
+   (e.g. a traceId embedded in the message) from overflowing the card. */
+.data-table__card .shipment-status-cell__error {
+  max-width: 100%;
+}
+
+.data-table__card .shipment-status-cell__error-message {
+  white-space: normal;
+  overflow: visible;
+  text-overflow: clip;
+  overflow-wrap: anywhere;
+}
+
 /* Loading-state skeleton (avoids CLS on first paint while connections
  * settle — tech-review SUGGESTION). Reserves the same vertical footprint
  * the populated panel will occupy. */
